Forge Area and Shop
I was trying to finish the texture for my forge but everything broke. Decided not to waste anymore time so I found a forge asset along with some other really cool assets, like the candles. Added fire light effect to the forge and the light sources in the shop/house.