Lots of Stuff
Alright I did a lot of stuff.
Fixed the lighting and the shadows are not pitch black anymore.
Edited the terrain, removed a lot of the unnecessary trees.
Fixed player size to fit through doorways
Added a fire effect to the forge
Added a smoke effect to the cooling bucket when the sword is dropped in. Also with this script the final product of the sword just kind of shoots out shortly after the smoke finishes. I think it's kind of funny.
Changed the shovel to be a pickax
Added a spark effect when the pickax hits the ore deposit
Changed the mining to produce 1 copper after being hit twice
Now when you get close enough to the counter in the shop a UI pop-up appears that says to press E to get a new order and it will make that new order when you press E
Rotate the tongs forward a little to make it look more realistic and fixed the way things attached to them when picked up (also disabled physics when holding something or else it would fling everywhere if it touched anything else).
Fixed the forge spawning multiple swords after putting one copper on it.
Added a basic cursor
Simplified the hitDamage script that checked collision with ore deposit, forge, and cooling bucket to 3 separate scripts. All being in one is what was causing the multiple swords to appear it was weird. There are 3 now, Mining, Melting and Cooling.